organize webpack plugins and setup sw update
