+++ config/webpack/plugins/html.js
