+++ config/webpack/plugins/manifest.js
