+++ config/webpack/plugins/offline.js
@@ -1,0 +1,7 @@
+const OfflinePlugin = require('offline-plugin');
+
+module.exports = new OfflinePlugin({
+  ServiceWorker: {
+    events: true
+  }
+});
